Fix erroneous torrent complete notifications when transmission restarts
When transmission restarts it will reassign ids to torrents. This means
that a downloading torrent with id x may still be downloading after
restart but the id is now assigned to a complete torrent which could
trigger a notification in error.

Now with name check it's not a complete fix but reduces the chance
of such match by a lot. Checking torrent hash would be best option
but it would increase traffic.